Changed all x/y values in gameobject related classes to float type

+++ phase1/app/src/main/java/com/example/phase1/Character.java
@@ -7,7 +7,7 @@ public class Character extends GameObject {
     private int health;
 
 
-    public Character(int x, int y) {  //NEED GameManager.getHealth()
+    public Character(float x, float y) {  //NEED GameManager.getHealth()
         super(x, y);
         this.speed = 80; //temp value of speed
         this.strength = 1;// temp
